# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "kolenke",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/settings/resume": {
      "post": {
        "tags": [
          "settings"
        ],
        "summary": "Upload Resume",
        "operationId": "upload_resume",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_resume"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResumeUploaded"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/resume-reviews": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "List Reviews",
        "operationId": "list_reviews",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReviewSummary"
                  },
                  "type": "array",
                  "title": "Response List Reviews"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Create Review",
        "description": "One of: a file (PDF, DOCX, TXT), pasted text, or use_saved_file for the resume uploaded in the settings.",
        "operationId": "create_review",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_review"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume-reviews/{rid}": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Get Review",
        "operationId": "get_review",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "resume"
        ],
        "summary": "Delete Review",
        "operationId": "delete_review",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume-reviews/{rid}/export": {
      "get": {
        "tags": [
          "resume"
        ],
        "summary": "Export Review",
        "operationId": "export_review",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The report as a Markdown file",
            "content": {
              "text/markdown": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resume-reviews/{rid}/rerun": {
      "post": {
        "tags": [
          "resume"
        ],
        "summary": "Rerun Review",
        "description": "The same resume, role and grade again: after the market moved, or to see the effect of new settings.",
        "operationId": "rerun_review",
        "parameters": [
          {
            "name": "rid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rid"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AiPart": {
        "properties": {
          "used": {
            "type": "boolean",
            "title": "Used",
            "default": false
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "default": ""
          },
          "grade_fit": {
            "type": "string",
            "title": "Grade Fit",
            "default": ""
          },
          "grade_comment": {
            "type": "string",
            "title": "Grade Comment",
            "default": ""
          },
          "error": {
            "type": "string",
            "title": "Error",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "used",
          "model",
          "summary",
          "grade_fit",
          "grade_comment",
          "error"
        ],
        "title": "AiPart"
      },
...
      "Body_create_review": {
        "properties": {
          "role": {
            "type": "string",
            "maxLength": 120,
            "minLength": 2,
            "title": "Role"
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "use_market": {
            "type": "boolean",
            "title": "Use Market",
            "default": true
          },
          "use_ai": {
            "type": "boolean",
            "title": "Use Ai",
            "default": false
          },
          "use_saved_file": {
            "type": "boolean",
            "title": "Use Saved File",
            "default": false
          },
          "text": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60000
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "file": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "role",
          "grade"
        ],
        "title": "Body_create_review"
      },
...
      "Body_upload_resume": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_resume"
      },
...
      "Check": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "group": {
            "type": "string",
            "title": "Group"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "$ref": "#/components/schemas/CheckStatus"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "fix": {
            "type": "string",
            "title": "Fix",
            "default": ""
          },
          "weight": {
            "type": "integer",
            "title": "Weight",
            "default": 1
          }
        },
        "type": "object",
        "required": [
          "id",
          "group",
          "title",
          "status",
          "detail",
          "fix",
          "weight"
        ],
        "title": "Check",
        "description": "One test the resume went through, e.g. «Достижения в цифрах»."
      },
      "CheckStatus": {
        "type": "string",
        "enum": [
          "pass",
          "warn",
          "fail"
        ],
        "title": "CheckStatus"
      },
...
      "Diff": {
        "properties": {
          "previous_id": {
            "type": "integer",
            "title": "Previous Id"
          },
          "previous_score": {
            "type": "integer",
            "title": "Previous Score"
          },
          "score_delta": {
            "type": "integer",
            "title": "Score Delta"
          },
          "fixed": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Fixed",
            "default": []
          },
          "new_issues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "New Issues",
            "default": []
          },
          "skills_added": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Added",
            "default": []
          },
          "same_resume": {
            "type": "boolean",
            "title": "Same Resume",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "previous_id",
          "previous_score",
          "score_delta",
          "fixed",
          "new_issues",
          "skills_added",
          "same_resume"
        ],
        "title": "Diff",
        "description": "Compared with the previous review of the same role and grade: what you already improved."
      },
...
      "Grade": {
        "type": "string",
        "enum": [
          "intern",
          "junior",
          "middle",
          "senior",
          "lead"
        ],
        "title": "Grade"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Market": {
        "properties": {
          "query": {
            "type": "string",
            "title": "Query"
          },
          "sample_size": {
            "type": "integer",
            "title": "Sample Size"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources",
            "default": []
          },
          "note": {
            "type": "string",
            "title": "Note",
            "default": ""
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/MarketSkill"
            },
            "type": "array",
            "title": "Skills",
            "default": []
          },
          "coverage": {
            "type": "integer",
            "title": "Coverage",
            "default": 0
          },
          "salary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Salary"
              },
              {
                "type": "null"
              }
            ]
          },
          "title_words": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Title Words",
            "default": []
          },
          "examples": {
            "items": {
              "$ref": "#/components/schemas/MarketVacancyRef"
            },
            "type": "array",
            "title": "Examples",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "query",
          "sample_size",
          "sources",
          "note",
          "skills",
          "coverage",
          "salary",
          "title_words",
          "examples"
        ],
        "title": "Market"
      },
      "MarketSkill": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "share": {
            "type": "integer",
            "title": "Share"
          },
          "have": {
            "type": "boolean",
            "title": "Have"
          }
        },
        "type": "object",
        "required": [
          "name",
          "share",
          "have"
        ],
        "title": "MarketSkill"
      },
      "MarketVacancyRef": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company",
            "default": ""
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "title",
          "company",
          "url",
          "source"
        ],
        "title": "MarketVacancyRef"
      },
      "Ok": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "Ok"
      },
...
      "Recommendation": {
        "properties": {
          "priority": {
            "type": "integer",
            "title": "Priority"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "priority",
          "title",
          "detail"
        ],
        "title": "Recommendation"
      },
...
      "ResumeFacts": {
        "properties": {
          "words": {
            "type": "integer",
            "title": "Words"
          },
          "experience_months": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Months"
          },
          "contacts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contacts",
            "default": []
          },
          "skills_found": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Skills Found",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "words",
          "experience_months",
          "contacts",
          "skills_found"
        ],
        "title": "ResumeFacts"
      },
      "ResumeReport": {
        "properties": {
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "verdict": {
            "type": "string",
            "title": "Verdict"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "role_profile": {
            "type": "string",
            "title": "Role Profile",
            "default": ""
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "facts": {
            "$ref": "#/components/schemas/ResumeFacts"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/Check"
            },
            "type": "array",
            "title": "Checks"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Strengths"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Weaknesses"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "type": "array",
            "title": "Recommendations"
          },
          "rewrites": {
            "items": {
              "$ref": "#/components/schemas/Rewrite"
            },
            "type": "array",
            "title": "Rewrites"
          },
          "market": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Market"
              },
              {
                "type": "null"
              }
            ]
          },
          "ai": {
            "$ref": "#/components/schemas/AiPart",
            "default": {
              "used": false,
              "model": "",
              "summary": "",
              "grade_fit": "",
              "grade_comment": "",
              "error": ""
            }
          },
          "diff": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Diff"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "score",
          "verdict",
          "role",
          "role_profile",
          "grade",
          "facts",
          "checks",
          "strengths",
          "weaknesses",
          "recommendations",
          "rewrites",
          "market",
          "ai",
          "diff"
        ],
        "title": "ResumeReport"
      },
      "ResumeUploaded": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "name"
        ],
        "title": "ResumeUploaded"
      },
      "ReviewDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "progress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "use_market": {
            "type": "boolean",
            "title": "Use Market"
          },
          "use_ai": {
            "type": "boolean",
            "title": "Use Ai"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "report": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResumeReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "finished_at",
          "status",
          "progress",
          "role",
          "grade",
          "source_name",
          "use_market",
          "use_ai",
          "score",
          "error",
          "report"
        ],
        "title": "ReviewDetail"
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "done",
          "error"
        ],
        "title": "ReviewStatus"
      },
      "ReviewSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "progress": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "grade": {
            "$ref": "#/components/schemas/Grade"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "use_market": {
            "type": "boolean",
            "title": "Use Market"
          },
          "use_ai": {
            "type": "boolean",
            "title": "Use Ai"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created_at",
          "finished_at",
          "status",
          "progress",
          "role",
          "grade",
          "source_name",
          "use_market",
          "use_ai",
          "score",
          "error"
        ],
        "title": "ReviewSummary"
      },
      "Rewrite": {
        "properties": {
          "before": {
            "type": "string",
            "title": "Before"
          },
          "after": {
            "type": "string",
            "title": "After"
          },
          "why": {
            "type": "string",
            "title": "Why"
          },
          "by_ai": {
            "type": "boolean",
            "title": "By Ai",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "before",
          "after",
          "why",
          "by_ai"
        ],
        "title": "Rewrite"
      },
...
      "Salary": {
        "properties": {
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "low": {
            "type": "integer",
            "title": "Low"
          },
          "median": {
            "type": "integer",
            "title": "Median"
          },
          "high": {
            "type": "integer",
            "title": "High"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "currency",
          "low",
          "median",
          "high",
          "count"
        ],
        "title": "Salary"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}